pyproject.toml
[tool.poetry]
name = "py-finances-backend"
version = "0.1.0"
description = ""
authors = ["EDUARDO LIMA <weduardow123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
taskipy = "^1.12.2"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
openai = "^1.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
app = "fastapi dev src/app/main.py"